Update Recordings page (part 1)
1. refactor UI
2. update link
